requirements.txt
flask==3.0.0
line-bot-sdk==3.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
icalendar==5.0.11
pytz==2024.1
gunicorn==21.2.0
